Add seemless macos support and dialogue device selection
Previously, I had to remove the attributes from the .env every time, i changed that in openwhoop/src/main.rs so that it automatically macos and removes bac attributes (BLE_INTERFACE, WHOOP_ADDR). Also I added a cli-dialogue for decice selection which scans for 10 seconds all devices that offer the whoop service and stores it with their advertising names and addresses. Then you can choose your device directly in the cli. On macos every address is 00:00:00:00:00:00. So here we cant use the WHOOP_ADDR or the adversting BLE address that is shown. Instead we choose the detected device in the cli